Refactor worker state & IPC; improve UI & job handling
- Add WorkerStateChanged message and event-driven state tracking
- Use long IDs for jobs; unify message sending logic
- Implement lightweight (header-only) IPC messages
- Add async handshake and explicit ready signals
- Show worker states and progress in UI with color coding
- Refactor SemaphoreSlimExtensions for async/await
- Improve error handling, logging, and code clarity
- Add Hexa.NET.Utilities package reference
- Various bug fixes and protocol cleanups

+++ WorkerShared/Workload.cs
@@ -5,7 +5,7 @@
     [ProtobufRecord]
     public partial struct JobPayload : IRecord
     {
-        public int Id;
+        public long Id;
         public string Source;
         public string Destination;
         public WorkloadFlags Flags;
@@ -13,7 +13,7 @@
         public int MinSize;
         public int MaxSize;
 
-        public JobPayload(int id, string source, string destination, WorkloadFlags flags, int format, int minSize, int maxSize)
+        public JobPayload(long id, string source, string destination, WorkloadFlags flags, int format, int minSize, int maxSize)
         {
             Id = id;
             Source = source;
